mcp-server/pyproject.toml
[tool.poetry]
name = "figmaflow-mcp-server"
version = "0.1.0"
description = "MCP server for converting Figma designs to Flutter code"
authors = ["FigmaFlow Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
mcp = "^1.0.0"
httpx = "^0.27.0"
pydantic = "^2.0.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
ruff = "^0.3.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
